+++ src/core/editor/menu.rs
@@ -1,20 +1,20 @@
 use egui::Ui;
 
-use crate::core::editor::objects::editor_settings::EditorSettings;
+use crate::core::editor::objects::{settings::EditorSettings, ui_tree::PaneKind};
 
 
-pub(crate) struct EditorMenu{
+pub(crate) struct EditorMenu{}
 
-}
 impl EditorMenu{
-    ///NOTE: Return parameter is a `pane id` the change came from.
-    pub(crate) fn ui(&mut self, ui: &mut Ui, settings: &mut EditorSettings){
+    ///Collects `PaneKinds` of which settings have been altered.
+    pub(crate) fn ui(&mut self, ui: &mut Ui, settings: &mut EditorSettings) -> Option<Vec<PaneKind>>{
+        let mut ui_changes = vec![];
+
         egui::menu::bar(ui, |ui| {
             ui.menu_button("Settings", |ui| {
                 ui.menu_button("File Tree", |ui| {
-                    //Note: Value is automatically changed. No need to change it manually
                     if ui.checkbox(&mut settings.show_hidden_elements, "Show Hidden items").clicked(){
-                        //
+                        ui_changes.push(PaneKind::FileTree);
                     }
                     if ui.button("Show Hidden items").clicked() {
                         // handle action
@@ -29,5 +29,7 @@ impl EditorMenu{
                 }
             });
         });
+
+        ui_changes.into()
     }
 }