+++ src/core/editor/objects/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod flat_tree;
-pub(crate) mod editor_settings;
+pub(crate) mod settings;
+pub(crate) mod ui_tree;