+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 
 
+[0.1.12] - 19/7/2025
+Implemented settings inheritance from *EditorMenu -> EditorLayout*
+
+### Added 
+    - Live settings for UI components.
+### Changed
+    - Pane's are more self contained (added settings and reloading)
+
+
 [0.1.11] - 18/7/2025
 Rethinking relation between editor and ui components in regards to setting inheritance.